LayoutTests/fast/forms/state-restore-to-non-edited-controls-expected.txt

Test to NOT save state for non-edited controls

PASS document.getElementById("button").value is "2"
PASS document.getElementById("hidden").value is "2"
PASS document.getElementById("image").value is "2"
PASS document.getElementById("reset").value is "2"
PASS document.getElementById("submit1").value is "2"
PASS document.getElementById("text0").value is "2"
PASS document.getElementById("text1").value is "edit"
PASS document.getElementById("text2").value is "2"
PASS document.getElementById("textarea").value is "2"
